Move fetchCID into audius backend
