--- packages/web/src/services/audius-backend/index.ts
@@ -1,1 +1,0 @@
-export * from './fetchCID'
